refactor(core): extract shared OAuth flow primitives from MCPOAuthProvider
Extract generic OAuth 2.0 Authorization Code + PKCE logic into a shared
oauth-flow.ts module so both MCP and A2A authentication providers can
reuse it. MCPOAuthProvider now delegates to these shared utilities while
retaining MCP-specific logic (discovery, dynamic registration, resource
parameter handling).

Part of #17600
